% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MIC.R
\name{MIC}
\alias{MIC}
\title{MIC}
\usage{
MIC(cache.ok = TRUE, verbose = FALSE)
}
\arguments{
\item{cache.ok}{TRUE by default. If FALSE, the file will be redownloaded}

\item{verbose}{FALSE by default. Do you want to see messages telling you
whether the cached values were used or not?}
}
\value{
data.table
}
\description{
Market Identifier Codes
}
\details{
Download a spreadsheet with Market Identifier Codes and coerce to data.table.
This is actually a closure.  The results are cached, so that if you call this
funciton twice, it will return the cached values the second time instead of
redownloading the file.
}
\examples{
\dontrun{
MIC() # downloads file
MIC() # recalls cached data.table
}
}
\author{
gsee
}

