% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/symbology-package.R
\docType{package}
\name{symbology}
\alias{symbology}
\alias{symbology-package}
\title{symbology package: \R functions to access symbol mapping files available on
the internet.}
\description{
symbology package: \R functions to access symbol mapping files available on
the internet.
}
\examples{
\dontrun{
ArcaSymbolMap() # downloads from internet
ArcaSymbolMap() # fast. retrieves from memory
ArcaSymbolMap(cache.ok=FALSE) # re-downloads

ArcaEdgeSymbolMap()
NyseSymbolMap()
NyseGroupSymbols()
OpenBookSymbolMap()

BatsSymbolMap()

nasdaqtraded()
nasdaqlisted()
otherlisted()
otclist()
otcbb()

shortSaleRuleStocks()

mpidlist()
MIC()

NYSEHolidays()
}
}
\author{
Garrett See \email{gsee000@gmail.com}
}
\references{
\url{ftp://ftp.nysedata.com}
\url{http://www.batstrading.com/market_data/symbol_listing/csv/}
\url{ftp://ftp.nasdaqtrader.com}
\url{http://www.otcmarkets.com/reports/symbol_info.csv}
}
\keyword{IO}
\keyword{package}

